+++ src/mixer/mixer.h
@@ -16,8 +16,7 @@ class Mixer : public Model {
   // description: a short identifier for this mixer.
   Mixer(ShortTermMemory& short_term_memory, LongTermMemory& long_term_memory,
         unsigned int& context, float learning_rate, int layer_number,
-        unsigned long long table_size, std::string description,
-        bool enable_analysis);
+        unsigned int table_size, std::string description, bool enable_analysis);
   void Predict(ShortTermMemory& short_term_memory,
                const LongTermMemory& long_term_memory);
   void Learn(const ShortTermMemory& short_term_memory,
